+++ data-access-service-axon/src/test/java/uk/gov/justice/laa/dstew/access/query/application/priorauthority/PriorAuthorityProjectionTest.java
@@ -326,7 +326,7 @@ class PriorAuthorityProjectionTest {
                 new PriorAuthorityDataPayload.DecisionDetails(
                     "GRANTED",
                     "Decision recorded",
-                    99.99,
+                    BigDecimal.valueOf(99.99),
                     Instant.parse("2026-09-08T12:00:00Z"),
                     null,
                     DisbursementInformation.builder().build(),
